+++ logs.js
@@ -1,4 +1,4 @@
-export const logs = {
+export const LOGS = {
   start:
     'Часы показывали [time], когда [player1] и [player2] бросили вызов друг другу.',
   end: [
