+++ playerHp.js
@@ -1,22 +1,22 @@
-export const attack = () => {
-  console.log(`${this.name}${' Fight!!!'}`)
-}
+// export const attack = () => {
+//   console.log(`${this.name}${' Fight!!!'}`)
+// }
 
-export const getRandom = (num) => {
-  return Math.ceil(Math.random() * num)
-}
+// export const getRandom = (num) => {
+//   return Math.ceil(Math.random() * num)
+// }
 
-export function elHP() {
-  return document.querySelector(`.player${this.player} .life`)
-}
+// export function elHP() {
+//   return document.querySelector(`.player${this.player} .life`)
+// }
 
-export function changeHp(changeHP) {
-  this.hp -= getRandom(changeHP)
-  if (this.hp <= 0) {
-    this.hp = 0
-  }
-}
+// export function changeHp(changeHP) {
+//   this.hp -= getRandom(changeHP)
+//   if (this.hp <= 0) {
+//     this.hp = 0
+//   }
+// }
 
-export function renderHP() {
-  this.elHP().style.width = `${this.hp}${'%'}`
-}
+// export function renderHP() {
+//   this.elHP().style.width = `${this.hp}${'%'}`
+// }
