+++ utils.js
@@ -1,0 +1,3 @@
+export const getRandom = (num) => {
+  return Math.ceil(Math.random() * num)
+}
